update typedefs &resolvers
