+++ client/src/components/Card.js
@@ -57,19 +57,5 @@ const Card = ({ image, author, title, description, id, price }) => {
         </div>
     )
 };
-function fetchdata() {
-    var data = { "price_ID": "price_1M4whsJL7p827Ta0DKuJFcBR" };
-    console.log(data, "data");
-    fetch('/create-checkout-session',
-        {
-            method: "POST",
-            body: JSON.stringify(data),
-            headers: {
-                "Content-Type": "application/json"
-            }
-        })
-        .then((res) => res.json())
 
-        .then((data) => console.log(data, "data2"))
-};
 export default Card;
