+++ client/src/components/checkout.js
@@ -1,41 +1,42 @@
 import React, { useState, useEffect } from "react";
+import Card from "./Card";
 //import "./App.css";
 
 
 
-const ProductDisplay = () => (
-    <section>
-        <div className="product">
-            <img
-                src="https://i.imgur.com/EHyR2nP.png"
-                alt="The cover of Stubborn Attachments"
-            />
-            <div className="description">
-                <h3>Stubborn Attachments</h3>
-                <h5>$20.00</h5>
-            </div>
-        </div>
-        <div action="http://localhost:3001/create-checkout-session" >
-            <button onClick={fetchdata}>
-                Checkout
-            </button>
-        </div>
-    </section >
-);
+// const ProductDisplay = () => (
+//     <section>
+//         <div className="product">
+//             <img
+//                 src="https://i.imgur.com/EHyR2nP.png"
+//                 alt="The cover of Stubborn Attachments"
+//             />
+//             <div className="description">
+//                 <h3>Stubborn Attachments</h3>
+//                 <h5>$20.00</h5>
+//             </div>
+//         </div>
+//         <div action="http://localhost:3001/create-checkout-session" >
+//             <button onClick={fetchdata}>
+//                 Checkout
+//             </button>
+//         </div>
+//     </section >
+// );
 
-function fetchdata() {
-    var data = { "price_ID": "price_1M4whsJL7p827Ta0DKuJFcBR" };
+// function fetchdata() {
+//     var data = { "price_ID": "price_1M4whsJL7p827Ta0DKuJFcBR" };
 
-    fetch('/create-checkout-session', {
-        method: "POST",
-        body: JSON.stringify(data),
-        headers: {
-            "Content-Type": "application/json"
-        }
-    })
-        .then((res) => res.json())
-        .then((data) => console.log(data, "data"));
-}
+//     fetch('/create-checkout-session', {
+//         method: "POST",
+//         body: JSON.stringify(data),
+//         headers: {
+//             "Content-Type": "application/json"
+//         }
+//     })
+//         .then((res) => res.json())
+//         .then((data) => console.log(data, "data"));
+// }
 const Message = ({ message }) => (
     <section>
         <p>{message}</p>
@@ -65,6 +66,6 @@ export default function Checkout() {
     return message ? (
         <Message message={message} />
     ) : (
-        <ProductDisplay />
+        <Card />
     );
 }
