+++ server/schemas/typeDefs.js
@@ -16,7 +16,6 @@ const typeDefs = gql`
         soldItems: [Item]
         soldCount: Int
     }
-
     type Item {
         _id: ID
         author: String
@@ -25,12 +24,10 @@ const typeDefs = gql`
         price: Float
         image: String        
     }
-
     type Auth {
         token: ID!
         user: User
     }
-
     type Query {
         me: User
         items: [Item]
@@ -39,22 +36,18 @@ const typeDefs = gql`
         #item por tag || descr || title || author
         
     }
-
     type Mutation {
         addUser(username: String!, email: String!, password: String!): Auth        
         login(email: String!, password: String!): Auth
-
         updateUser(street: String!, zip: String!, city: String!, state: String!, country: String!): User
         
         addItemToSale(author: String!, title: String!, description: String!, price: Float!, image: String): Item
         deleteItemToSale(itemId: ID!): Item
         updateItemToSale(author: String!, title: String!, description: String!, price: Float!, image: String): Item
-
         addBoughtItem(itemId: ID!): Item
         addSoldItem(itemId: ID!): Item
         addToCart(itemId: ID!): User
         deleteFromCart(itemId: ID!): User
-
     }
 `;
